Move WindowEvent, bcd_to_dec and dec_to_bcd to utils
