+++ extras/examples/gamewrapper_tetris.py
@@ -10,7 +10,8 @@ import sys
 file_path = os.path.dirname(os.path.realpath(__file__))
 sys.path.insert(0, file_path + "/../..")
 
-from pyboy import PyBoy, WindowEvent # isort:skip
+from pyboy import PyBoy # isort:skip
+from pyboy.utils import WindowEvent # isort:skip
 
 # Check if the ROM is given through argv
 if len(sys.argv) > 1:
