+++ pyboy/__init__.py
@@ -7,11 +7,9 @@ __pdoc__ = {
     "core": False,
     "logging": False,
     "pyboy": False,
-    "utils": False,
     "conftest": False,
 }
 
-__all__ = ["PyBoy", "PyBoyMemoryView", "WindowEvent", "dec_to_bcd", "bcd_to_dec"]
+__all__ = ["PyBoy", "PyBoyMemoryView"]
 
 from .pyboy import PyBoy, PyBoyMemoryView
-from .utils import WindowEvent, bcd_to_dec, dec_to_bcd
